requirements.txt
discord.py==1.7.3
python-dotenv==0.19.0
psycopg2==2.9.3
requests==2.19.1